refactor(stm): moved the ensure_advanceable_rolling_state function to the rolling state struct
